refactor(FixedCardsShuffler): 패키지 이동
