+++ src/test/java/domain/shuffler/FixedCardsShuffler.java
@@ -1,6 +1,7 @@
-package domain.card.shuffler;
+package domain.shuffler;
 
 import domain.card.Card;
+import domain.card.shuffler.CardsShuffler;
 import java.util.Deque;
 import java.util.Stack;
